chore: Update form button styles and add login button to Navbar
Update the form button styles in MainPage.scss and Links.jsx to adjust the width and font size. Also, add a login button to the Navbar component in Navbar.jsx.

+++ src/pages/Profile.jsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from "react"
-import "./MainPage.scss"
 import { useNavigate, useParams } from "react-router-dom"
 import Navbar from "./navbar/Navbar"
 import Section from "./sections/Section"
@@ -10,6 +9,7 @@ import { LuMousePointerClick } from "react-icons/lu";
 import AlertModal from "./modals/AlertModal";
 import { useRecoilState } from 'recoil';
 import { userState } from './user_session_state';
+import "../pages/portfolio/Portfolio.scss";
 
 const Profile = () => {
     const [user_id, setUserState] = useRecoilState(userState);
@@ -21,6 +21,7 @@ const Profile = () => {
     const [confirmPassword, setConfirmPassword] = useState('');    
     const [showAlertModal, setShowAlertModal] = useState(false);
     const [showAlertModalPassword, setShowAlertModalPassword] = useState(false);
+    const [showAlertUserExists, setShowAlertUserExists] = useState(false);  
 
     useEffect(() => {
         fetch(`/profile/${user_id}`, {
@@ -62,13 +63,20 @@ const Profile = () => {
             }),
         });
         const data = await response.json();
+        if (data === "User already exists") {
+            setShowAlertUserExists(true);
+            return;
+        }
+
         setUserState(data);
         window.localStorage.setItem('user_id', data);
     };
 
-    if (user_id !== 'undefined') {
+    console.log(user_id);
+
+    if (!user_id) {
         return (
-            <div className="main-body">
+            <div className="portfolio-body">
                 <Navbar />
                 <div className="main">
                     <Section title="My Profile" />
@@ -114,10 +122,11 @@ const Profile = () => {
         );
     } else {
         return (
-            <div className="main-body">
+            <div className="portfolio-body">
                 <Navbar />
                 <Section title="Create an Account" />
-                <div className="main">
+                <br/>
+                <div className="portfolio">
                     <AlertModal
                         modal={showAlertModal}
                         toggle={() => setShowAlertModal(false)}
@@ -129,6 +138,12 @@ const Profile = () => {
                         toggle={() => setShowAlertModalPassword(false)}
                         message="Passwords do not match"
                         isOpen={showAlertModalPassword}
+                    />
+                    <AlertModal
+                        modal={showAlertUserExists}
+                        toggle={() => setShowAlertUserExists(false)}
+                        message="User already exists"
+                        isOpen={showAlertUserExists}
                     />
                     <Form onSubmit={handleSubmit}>
                         <FormGroup row>
